requirements.txt
pandas>=3.0.0
jupyter>=1.0.0
fastapi>=0.128.2
azure-storage-blob>=12.19.0
requests>=2.31.0
python-dotenv>=1.0.0
ipython-sql>=0.5.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0